Cleanup and upgrade
