+++ MozuAndroidInStoreAssistant/app/build.gradle
@@ -26,12 +26,7 @@ android {
         minSdkVersion 14
         targetSdkVersion 19
         versionCode 102
-        versionName "1.0.2 - "+getDate()+" - Build " + System.getenv("BUILD_NUMBER")
-        testApplicationId "com.mozu.mozuandroidinstoreassistant.app.test"
-        testInstrumentationRunner 'com.google.android.apps.common.testing.testrunner.GoogleInstrumentationTestRunner'
-        testHandleProfiling true
-        testFunctionalTest true
-        testApplicationId 'com.mozu.mozuandroidinstoreassistant.app.test'
+        versionName "1.0.2 - " + getDate() + " - Build " + System.getenv("BUILD_NUMBER")
     }
     signingConfigs {
         release {
@@ -41,79 +36,51 @@ android {
             keyAlias releaseKeyAlias
             keyPassword releaseKeyPassword
         }
-        debug {
-            storeFile file('debug.keystore')
-            storePassword 'android'
-            keyAlias 'androiddebugkey'
-            keyPassword 'android'
-        }
     }
     buildTypes {
         release {
             minifyEnabled false
             signingConfig signingConfigs.release
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.txt'
-            ext.betaDistributionEmails="shadaab_kazi@volusion.com, adam_dutton@volusion.com, santhosh_mankala@volusion.com, chris_pound@volusion.com, rabin_joshi@volusion.com, volusionmozu@gmail.com"
-            ext.betaDistributionReleaseNotesFilePath=file('releasenotes.txt')
-            ext.enableCrashlytics = enableCrashlytics.toBoolean()
-            buildConfigField "String", "APP_AUTH_APPID","\""+mozuAuthAppId+"\""
-            buildConfigField "String", "APP_AUTH_SHARED_SECRET","\""+mozuAuthSharedSecret+"\""
-            buildConfigField "String", "SERVICE_URL","\""+mozuServiceURL+"\""
-            if(enableCrashlytics.toBoolean()) {
-                buildConfigField "boolean", "ENABLE_CRASHLYTICS", "true"
-            }else{
-                buildConfigField "boolean", "ENABLE_CRASHLYTICS", "false"
-            }
         }
         debug {
             minifyEnabled false
             signingConfig signingConfigs.debug
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.txt'
-            ext.betaDistributionEmails="shadaab_kazi@volusion.com, adam_dutton@volusion.com, santhosh_mankala@volusion.com, chris_pound@volusion.com, rabin_joshi@volusion.com, volusionmozu@gmail.com"
-            ext.betaDistributionReleaseNotesFilePath=file('releasenotes.txt')
-            ext.enableCrashlytics = enableCrashlytics.toBoolean()
-            buildConfigField "String", "APP_AUTH_APPID","\""+mozuAuthAppId+"\""
-            buildConfigField "String", "APP_AUTH_SHARED_SECRET","\""+mozuAuthSharedSecret+"\""
-            buildConfigField "String", "SERVICE_URL","\""+mozuServiceURL+"\""
-            if(enableCrashlytics.toBoolean()) {
-                buildConfigField "boolean", "ENABLE_CRASHLYTICS", "true"
-            }else{
-                buildConfigField "boolean", "ENABLE_CRASHLYTICS", "false"
-            }
         }
 
         applicationVariants.all { variant ->
-            if (variant.name == "debug"  && enableCrashlytics.toBoolean()) {
+            if (enableCrashlytics.toBoolean()) {
                 File crashlyticsProperties = new File("${project.projectDir.absolutePath}/crashlytics.properties")
                 Properties properties = new Properties()
-                properties.put("apiSecret",crashlyticsApiSecret)
+                properties.put("apiSecret", crashlyticsApiSecret)
                 properties.put("apiKey", crashlyticsApiKey)
                 properties.store(new FileWriter(crashlyticsProperties), "")
             }
+
+            ext.betaDistributionEmails = "shadaab_kazi@volusion.com, adam_dutton@volusion.com, santhosh_mankala@volusion.com, chris_pound@volusion.com, rabin_joshi@volusion.com, volusionmozu@gmail.com"
+            ext.betaDistributionReleaseNotesFilePath = file('releasenotes.txt')
+            ext.enableCrashlytics = enableCrashlytics.toBoolean()
+            buildConfigField "String", "APP_AUTH_APPID", "\"" + mozuAuthAppId + "\""
+            buildConfigField "String", "APP_AUTH_SHARED_SECRET", "\"" + mozuAuthSharedSecret + "\""
+            buildConfigField "String", "SERVICE_URL", "\"" + mozuServiceURL + "\""
+            if (enableCrashlytics.toBoolean()) {
+                buildConfigField "boolean", "ENABLE_CRASHLYTICS", "true"
+            } else {
+                buildConfigField "boolean", "ENABLE_CRASHLYTICS", "false"
+            }
         }
     }
-
-
-
     packagingOptions {
         exclude 'META-INF/LICENSE'
         exclude 'META-INF/NOTICE'
         exclude 'META-INF/LICENSE.txt'
         exclude 'META-INF/NOTICE.txt'
     }
-    sourceSets {
-        androidTest {
-            java.srcDirs = ['src/test/java']
-        }
-    }
-    testOptions {
-        reportDir = "$project.buildDir/mozu/tests/report"
-    }
-
 }
 
 dependencies {
-    compile(name:'viewpager', ext:'aar')
+    compile(name: 'viewpager', ext: 'aar')
     compile 'com.mozu:mozu-api-android:1.13.0'
     compile 'com.crashlytics.android:crashlytics:1.+'
     compile 'com.squareup.picasso:picasso:2.3.4'
@@ -122,8 +89,6 @@ dependencies {
     compile 'org.joda:joda-convert:1.6'
     compile 'com.jakewharton:butterknife:5.1.2'
     compile 'com.netflix.rxjava:rxjava-android:0.17.5'
-
-    androidTestCompile files('libs/espresso-1.1-bundled.jar')
 }
 
 def getDate() {
